Refactored the project to keep all values as int instead of float

+++ audio-video/lab/includes/Block88.h
@@ -18,8 +18,8 @@ public:
 	void printBlock();
 
 //private:
-	float m_values[8][8];
-	float m_subValues[4][4];
+	int m_values[8][8];
+	int m_subValues[4][4];
 	int m_positionX;
 	int m_positionY;
 };
